Move some color-related functions to a separate header
